cleanup: remove redundant casts
git-svn-id: https://svn.apache.org/repos/asf/wicket/trunk@1038654 13f79535-47bb-0310-9956-ffa450edef68

+++ wicket-request/src/main/java/org/apache/wicket/request/mapper/info/PageInfo.java
@@ -45,7 +45,7 @@ public class PageInfo
 	 */
 	public PageInfo()
 	{
-		this((Integer)null);
+		this(null);
 	}
 
 	/**
